Automatically detect versioned associations
Can still define versioned associations manually using `has_versioned_association` if versioned classes aren't loaded yet, or if associations are defined dynamically.

e.g. in the case of the test suite, classes are defined all in the same file, so rails's autoloading is unable to load the models on-demand. This means when we check to see if an association is versioned, unless the model for that association is loaded, it will explode. Being able to pass a list of versioned associations allows us to skip autodetection, and avoid loading each associated model to see if it "has_hindsight".