#new_version no longer affects un-versioned HM :dependent => :destroy associations on previous versions
An un-versioned has_many collection moves from version to version as new versions are saved. If the new version assigns a new collection, the records from the previous collection were being destroyed as they were unassigned from the association. However, there is no need to move the old collection to the new version if it is about to be overwritten by the attributes argument passed to the new version. Therefore, we assign the new attributes first, then skip copying any associations that become loaded as a result.